+++ src/main/java/br/pro/brand/finance/models/entity/User.java
@@ -21,7 +21,7 @@ public class User implements Serializable {
 
     @Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 
     private String name;
 
@@ -37,7 +37,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(Integer id, String name, String email, String password) {
+    public User(Long id, String name, String email, String password) {
         super();
         this.id = id;
         this.name = name;
@@ -46,11 +46,11 @@ public class User implements Serializable {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
